requirements.txt
streamlit
google-cloud-vision
pdfplumber
Pillow
google-generativeai
